Decompose CommandCalendar into grid and dialog siblings
Pull the 742-line CommandCalendar apart into a 242-line orchestrator
plus three siblings: CalendarGrid renders the month view and day cells,
DayDetailsDialog shows the selected day's schedule and override
choices, and OverrideReasonDialog captures the optional override reason
when switching schedules.

Refs #48
